+++ blocklets/ai-studio/api/src/routes/ws.ts
@@ -2,7 +2,8 @@ import { user } from '@blocklet/sdk/lib/middlewares';
 import express from 'express';
 
 import { ensurePromptsEditor, isRefReadOnly } from '../libs/security';
-import { getRepository, projects } from '../store/projects';
+import Projects from '../store/models/projects';
+import { getRepository } from '../store/projects';
 
 const router = express.Router();
 
@@ -15,7 +16,7 @@ router.use(ensurePromptsEditor, user()).ws('/ws/:projectId/:ref', async (conn, r
   const { role } = req.user!;
   const readOnly = isRefReadOnly({ ref, role });
 
-  const project = await projects.findOne({ _id: projectId });
+  const project = await Projects.findOne({ where: { _id: projectId } });
   if (!project) {
     conn.close(3001, `Project ${projectId} not found`);
     return;